requirements.txt
alembic==1.5.8
aniso8601==9.0.1
astroid==2.5.1
attrs==20.3.0
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==35.0.0
debugpy==1.5.1
dnspython==2.1.0
email-validator==1.1.3
Faker==6.6.2
flake8==4.0.1
Flask==1.1.2
Flask-Migrate==2.7.0
flask-restplus==0.13.0
flask-serialize==2.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
future==0.18.2
greenlet==1.0.0
idna==2.10
importlib-metadata==4.8.1
importlib-resources==5.3.0
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
lazy-object-proxy==1.6.0
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
permissive-dict==1.0.4
psycopg2==2.9.1
psycopg2-binary==2.8.6
pycodestyle==2.8.0
pycparser==2.20
pyflakes==2.4.0
PyJWT==2.3.0
pylint==2.7.2
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.3
requests==2.25.1
six==1.15.0
SQLAlchemy==1.4.2
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.4
uuid==1.30
Werkzeug==0.16.1
wrapt==1.12.1
WTForms==2.3.3
zipp==3.6.0
